requirements.txt
transformers
faiss-cpu
numpy
